finish deploy

+++ src/main/java/com/myexam/domain/service/RecipeServiceImpl.java
@@ -7,6 +7,7 @@ import com.myexam.domain.repositories.RecipeRepository;
 import com.myexam.domain.repositories.entity.RecipeEntity;
 import com.myexam.exception.RecipeNotFoundException;
 import com.myexam.model.Recipe;
+import jakarta.validation.Valid;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
@@ -14,8 +15,10 @@ import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
+import org.springframework.validation.annotation.Validated;
 
 @Service
+@Validated
 @Qualifier("RecipeServiceImpl")
 public class RecipeServiceImpl implements RecipeService {
 
@@ -38,7 +41,7 @@ public class RecipeServiceImpl implements RecipeService {
     entity.setUpdated_at(new Date());
     var result = repository.saveAndFlush(entity);
 
-    var newRecipe = new Recipe();
+    var newRecipe = Recipe.fromDatasource(entity);
     newRecipe.setId(result.getId());
     newRecipe.setTitle(result.getTitle());
     newRecipe.setServes(result.getServes());
@@ -58,6 +61,7 @@ public class RecipeServiceImpl implements RecipeService {
       throw new RecipeNotFoundException("Recipe id is not found");
     }
     var entity = existingRecipe.get();
+
     entity.setTitle(recipe.getTitle());
     entity.setMaking_time(recipe.getMaking_time());
     entity.setServes(recipe.getServes());
@@ -65,21 +69,20 @@ public class RecipeServiceImpl implements RecipeService {
     entity.setCost(recipe.getCost());
     entity.setUpdated_at(new Date());
     //    entity.setCreated_at(new Date()); // 作成日付は更新しない
-    var result = repository.saveAndFlush(entity);
 
-    var newRecipe = new Recipe();
-    newRecipe.setId(result.getId());
-    newRecipe.setTitle(result.getTitle());
-    newRecipe.setServes(result.getServes());
-    newRecipe.setMaking_time(result.getMaking_time());
-    newRecipe.setCost(result.getCost());
+    repository.saveAndFlush(entity);
+    var updatedRecipe = Recipe.fromDatasource(entity);
     List list = new ArrayList<RecipeEntity>();
-    list.add(newRecipe);
+    list.add(updatedRecipe);
     return list;
   }
 
   @Override
   public void delete(long id) {
+    Optional<RecipeEntity> existingRecipe = repository.findById(id);
+    if (existingRecipe.isEmpty()) {
+      throw new RecipeNotFoundException("Recipe id is not found");
+    }
     repository.deleteById(id);
   }
 
@@ -88,12 +91,14 @@ public class RecipeServiceImpl implements RecipeService {
     var result = repository.findAll();
     List<Recipe> list = new ArrayList();
     result.forEach(v -> {
-      var recipe = new Recipe();
-      recipe.setId(v.getId());
-      recipe.setTitle(v.getTitle());
-      recipe.setServes(v.getServes());
-      recipe.setMaking_time(v.getMaking_time());
-      recipe.setCost(v.getCost());
+      var recipe =Recipe.of(
+              v.getId(),
+              v.getTitle(),
+              v.getServes(),
+              v.getMaking_time(),
+              v.getIngredients(),
+              v.getCost()
+      );
       list.add(recipe);
     });
 
@@ -103,13 +108,19 @@ public class RecipeServiceImpl implements RecipeService {
 
   @Override
   public List<Recipe> getById(long id) {
-    var result = repository.getOne(id);
-    var recipe = new Recipe();
-    recipe.setId(result.getId());
-    recipe.setTitle(result.getTitle());
-    recipe.setServes(result.getServes());
-    recipe.setMaking_time(result.getMaking_time());
-    recipe.setCost(result.getCost());
+    Optional<RecipeEntity> existingRecipe = repository.findById(id);
+    if (existingRecipe.isEmpty()) {
+      throw new RecipeNotFoundException("Recipe id is not found");
+    }
+    var result = existingRecipe.get();
+    var recipe = Recipe.of(
+            result.getId(),
+            result.getTitle(),
+            result.getServes(),
+            result.getMaking_time(),
+            result.getIngredients(),
+            result.getCost()
+    );
     List list = new ArrayList<RecipeEntity>();
     list.add(recipe);
     return list;
